daemon: move prompting endpoints to /v2/interfaces/prompting/{requests,rules}
